refactor: 패키지 구조를 변경, 위치 이동
