+++ src/main/java/roomescape/config/InterceptorConfig.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.interceptor.AdminInterceptor;
+import roomescape.controller.interceptor.AdminInterceptor;
 
 @Configuration
 @RequiredArgsConstructor
