+++ src/main/java/roomescape/config/ResolverConfig.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.resolver.JWTBearerTokenSubjectResolver;
+import roomescape.controller.resolver.JWTBearerTokenSubjectResolver;
 
 import java.util.List;
