+++ src/main/java/roomescape/controller/LoginController.java
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.LoginControllerTokenPostBody;
-import roomescape.dto.LoginControllerTokenPostResponse;
+import roomescape.controller.dto.LoginControllerTokenPostBody;
+import roomescape.controller.dto.LoginControllerTokenPostResponse;
 import roomescape.service.LoginService;
 
 import javax.validation.Valid;
