+++ src/main/java/roomescape/controller/dto/ThemeControllerGetResponse.java
@@ -1,21 +1,23 @@
-package roomescape.dto;
+package roomescape.controller.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.PositiveOrZero;
 
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class ThemeControllerPostBody {
-    @NotBlank
+public class ThemeControllerGetResponse {
+    @NotNull
+    private Long id;
+
+    @NotNull
     private String name;
 
-    @NotBlank
+    @NotNull
     private String desc;
 
     @NotNull
