+++ src/main/java/roomescape/controller/dto/ThemeControllerPostResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.controller.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
