+++ src/main/java/roomescape/controller/resolver/JWTBearerTokenSubjectResolver.java
@@ -1,4 +1,4 @@
-package roomescape.resolver;
+package roomescape.controller.resolver;
 
 
 import lombok.RequiredArgsConstructor;
