+++ src/main/java/roomescape/service/LoginService.java
@@ -2,8 +2,8 @@ package roomescape.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import roomescape.controller.dto.LoginControllerTokenPostBody;
 import roomescape.controller.errors.ErrorCode;
-import roomescape.dto.LoginControllerTokenPostBody;
 import roomescape.repository.MemberRepository;
 import roomescape.service.exception.ServiceException;
