+++ src/main/java/roomescape/service/ReservationService.java
@@ -3,8 +3,8 @@ package roomescape.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PathVariable;
+import roomescape.controller.dto.ReservationsControllerPostBody;
 import roomescape.controller.errors.ErrorCode;
-import roomescape.dto.ReservationsControllerPostBody;
 import roomescape.entity.Reservation;
 import roomescape.repository.ReservationRepository;
 import roomescape.service.exception.ServiceException;
