+++ src/main/java/roomescape/service/ThemeService.java
@@ -3,8 +3,8 @@ package roomescape.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
+import roomescape.controller.dto.ThemeControllerPostBody;
 import roomescape.controller.errors.ErrorCode;
-import roomescape.dto.ThemeControllerPostBody;
 import roomescape.entity.Theme;
 import roomescape.repository.ThemeRepository;
 import roomescape.service.exception.ServiceException;
