+++ src/test/java/roomescape/controller/MemberLoginTest.java
@@ -12,10 +12,10 @@ import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.SpringWebApplication;
-import roomescape.dto.LoginControllerTokenPostBody;
-import roomescape.dto.LoginControllerTokenPostResponse;
-import roomescape.dto.MemberControllerPostBody;
-import roomescape.dto.MemberControllerPostResponse;
+import roomescape.controller.dto.LoginControllerTokenPostBody;
+import roomescape.controller.dto.LoginControllerTokenPostResponse;
+import roomescape.controller.dto.MemberControllerPostBody;
+import roomescape.controller.dto.MemberControllerPostResponse;
 import roomescape.repository.MemberRepository;
 import roomescape.service.JWTProvider;
